Rename example_peripheral to sensor_tag
