+++ example/lib/devices_list/devices_bloc.dart
@@ -6,7 +6,7 @@ import 'package:blemulator_example/repository/device_repository.dart';
 import 'package:rxdart/rxdart.dart';
 import 'package:flutter_ble_lib/flutter_ble_lib.dart';
 import 'package:blemulator/blemulator.dart';
-import 'package:blemulator_example/example_peripherals/example_peripheral.dart';
+import 'package:blemulator_example/example_peripherals/sensor_tag.dart';
 
 class DevicesBloc {
   final List<BleDevice> bleDevices = <BleDevice>[];
